services/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
redis==5.0.1
playwright==1.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1 